Updated the instructions for adding Provider, because it's deprecated.
"The store injection pattern was popularized by the mobx-react library. It refers to a now obsolete way of accessing the MobX stores across the component tree. It was introduced because the React legacy context was rather awkward to use.

We are in 2019 now and we don't need injections anymore (kids, don't do drugs)."

Ref: https://mobx-react.js.org/recipes-inject